server/requirements.txt
numpy~=2.0.1
torch~=2.4.0
uvicorn~=0.20.0
fastapi~=0.103.0
halo~=0.0.31
passlib~=1.7.4
pydantic~=2.5.3
slowapi~=0.1.9
sqlalchemy~=2.0.30
pyjwt~=2.8.0
cachetools~=5.4.0
python-dotenv~=1.0.1
httpx~=0.27.0
aiohttp~=3.9.5